requirements.txt
streamlit
python-dotenv
pinecone
langchain
langchain-community
langchain-text-splitters
langchain-groq
